+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/CoreModuleConfig.java
@@ -103,6 +103,18 @@ public class CoreModuleConfig extends ModuleConfig {
      */
     private int maxSizeOfAnalyzeProfileSnapshot = 12000;
     /**
+     * Analyze eBPF profiling data max duration(minute)
+     */
+    private int maxDurationOfAnalyzeEBPFProfiling = 10;
+    /**
+     * Query the eBPF Profiling data max duration(second) from database.
+     */
+    private int maxDurationOfQueryEBPFProfilingData = 30;
+    /**
+     * Thread Count of query the eBPF Profiling data.
+     */
+    private int maxThreadCountOfQueryEBPFProfilingData = Runtime.getRuntime().availableProcessors();
+    /**
      * Extra model column are the column defined by {@link ScopeDefaultColumn.DefinedByField#requireDynamicActive()} ==
      * true. These columns of model are not required logically in aggregation or further query, and it will cause more
      * load for memory, network of OAP and storage.